+++ pkg/provider/diff.go
@@ -37,12 +37,14 @@ type properties []string
 
 var forceNew = _groups{
 	"apps": _versions{
-		// NOTE: .spec.selector triggers a replacement in Deployment only AFTER v1beta1.
+		// NOTE: .spec.selector triggers a replacement in Deployment/Daemonset only AFTER v1beta1.
 		"v1beta1": _kinds{"StatefulSet": statefulSet},
 		"v1beta2": _kinds{
+			"Daemonset": daemonset,
 			"Deployment":  deployment,
 			"StatefulSet": statefulSet},
 		"v1": _kinds{
+			"Daemonset": daemonset,
 			"Deployment":  deployment,
 			"StatefulSet": statefulSet},
 	},
@@ -134,6 +136,10 @@ var core = _versions{
 			".spec.type",
 		},
 	},
+}
+
+var daemonset = properties{
+	".spec.selector",
 }
 
 var deployment = properties{